requirements.txt
python-dotenv
asyncpg
requests
httpx